get by hair working
getter for hair was type int instead of double lol

+++ Project01 - Super Squish Sisters/Character.h
@@ -31,7 +31,7 @@ public:
     virtual ~Character();
     
     virtual void print() const = 0;
-    int getHair(){return m_hair_length;}
+    double getHair(){return m_hair_length;}
     string getSpeed(){return m_speed;}
     string getRace(){return m_race;}
     string getName(){return m_name;}
